Version 0.0.2
Added a new Scene to test touch input and I got it working and made it
play a sound on jump.

There is a counter that should count how high the player is but at the
moment it looks tiny on my phone even though it looks fine in the
editor.