Update 1
Deleted all non-endurance scenes. Also removed all mobile scenes as
touch worked on the desktop scenes (wot zack?).

Made main menu and game mode select lead to the game.

Froze the game until jump is pressed and when  jump is pressed, game is
unfrozen and time is counted towards time spent in the air.

Collision is detected, need to make some sparks or something (jelly idk)
and then freeze the game and allow the player to restart and try again
or submit their score to a scoreboard etc.